build(cmake): add use_seh, have_nanosleep options
Added the SQLITE_USE_SEH compile-time option that enables Structured
Exception Handling on Windows while working with the memory-mapped shm
file that is part of WAL mode processing. This option is enabled by
default when building on Windows using Makefile.msc.

The VFS for unix now assumes that the nanosleep() system call is
available unless compiled with -DHAVE_NANOSLEEP=0.

Signed-off-by: John Janssen <71168603+jlmjanssen@users.noreply.github.com>