build(cmake): add strict_subtype option
Added the SQLITE_RESULT_SUBTYPE property for application-defined SQL
functions. All application defined SQL functions that invokes
sqlite3_result_subtype() must be registered with this new property.
Failure to do so might cause the call to sqlite3_result_subtype() to
behave as a no-op. Compile with -DSQLITE_STRICT_SUBTYPE=1 to cause an
SQL error to be raised if a function that is not SQLITE_RESULT_SUBTYPE
tries invokes sqlite3_result_subtype(). The use of
-DSQLITE_STRICT_SUBTYPE=1 is a recommended compile-time option for
every application that makes use of subtypes.

Signed-off-by: John Janssen <71168603+jlmjanssen@users.noreply.github.com>